requirements.txt
Flask==2.3.2
tensorflow==2.18.0
numpy==1.23.5
Pillow==9.4.0
Werkzeug==2.3.0
